moves styled components from App.js to Common, moving common styled components to individual directories in progress
