+++ src/components/Common/Content/index.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const Content = styled.div`
+  grid-column: 1/3;
+  grid-row: 2;
+`;
+export default Content;
