+++ src/components/Common/Footer/index.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const Footer = styled.div`
+  grid-column: 2;
+  grid-row: 4;
+  background: black;
+`;
+
+export default Footer;
