+++ src/components/Common/PrimaryButton/index.js
@@ -1,0 +1,24 @@
+import styled from "styled-components";
+
+const PrimaryButton = styled.button`
+  border: none;
+  background: transparent;
+  color: #25bea0;
+  padding: 0.5rem 1rem 0.5rem 1rem;
+  font-size: 0.875rem;
+  transition: all 0.2s ease;
+  cursor: pointer;
+  outline: none;
+  border-radius: 5px;
+
+  :active {
+    background: #25bea0;
+    color: #ffffff;
+  }
+
+  :hover {
+    color: #ffffff;
+  }
+`;
+
+export default PrimaryButton;
