+++ src/components/Common/index.js
@@ -1,39 +1,5 @@
 import styled from "styled-components";
 
-export const PrimaryButton = styled.button`
-  border: none;
-  background: transparent;
-  color: #25bea0;
-  padding: 0.5rem 1rem 0.5rem 1rem;
-  font-size: 0.875rem;
-  transition: all 0.2s ease;
-  cursor: pointer;
-  outline: none;
-  border-radius: 5px;
-
-  :active {
-    background: #25bea0;
-    color: #ffffff;
-  }
-
-  :hover {
-    color: #ffffff;
-  }
-`;
-
-export const Logo = styled.img`
-  order: 0;
-  display: block;
-  background: transparent;
-  grid-column: 1;
-  grid-row: 1;
-  justify-content: flex-start;
-`;
-
-export const LoginLogo = styled(Logo)`
-  margin-top: 20px;
-`;
-
 export const DeleteButton = styled.button`
   border: none;
   background-color: white;
@@ -155,35 +121,6 @@ export const Overpane = styled.div`
   z-index: 1100;
   transition: 250ms cubic-bezier(0.7, 0, 0.3, 1) transform;
   transform: translateY(${props => (props.overpane ? "100%" : "0")});
-`;
-
-export const SubmitButton = styled(PrimaryButton)`
-  width: 150px;
-  height: 30px;
-  color: white;
-  border-radius: 5px;
-  background: #25bea0;
-  border: 1px solid grey;
-  font-size: 16px;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-  border: none;
-  margin-left: 10px;
-`;
-
-export const NavLink = styled.p`
-  font-size: 1rem;
-  text-decoration: none;
-  color: #ffffff;
-  cursor: pointer;
-  margin: 0 10px 0 10px;
-  @media (max-width: 1024px) {
-    margin: 0 auto;
-    padding: 20px;
-  }
-  @media (max-width: 800px) {
-    margin: 0 auto;
-  }
 `;
 
 export const SpinnerWrapper = styled.div`
