+++ src/components/Header/index.js
@@ -2,7 +2,9 @@ import React, { Component, Fragment } from "react";
 
 import { connect } from "react-redux";
 import styled from "styled-components";
-import { PrimaryButton, BurgerButton, NavLink } from "../Common";
+import PrimaryButton from "../Common/PrimaryButton";
+import { BurgerButton } from "../Common/";
+import PLink from "../Common/PLink";
 import Stripe from "../Stripe";
 import { HashLink as Link } from "react-router-hash-link";
 import {
@@ -172,7 +174,7 @@ class Header extends Component {
                     DASHBOARD
                   </MenuLink>
 
-                  <NavLink
+                  <PLink
                     onClick={() => {
                       this.props.toggleMobileMenu(
                         !this.props.mobileMenu,
@@ -181,7 +183,7 @@ class Header extends Component {
                     }}
                   >
                     PREFERENCES
-                  </NavLink>
+                  </PLink>
                   {/* <Link to="/register">
                     <RegisterButton>REFER A FRIEND </RegisterButton>
                   </Link> */}
@@ -196,7 +198,7 @@ class Header extends Component {
                 </Fragment>
               ) : (
                 <Fragment>
-                  <NavLink
+                  <PLink
                     onClick={() => {
                       this.props.toggleMobileMenu(
                         !this.props.mobileMenu,
@@ -205,7 +207,7 @@ class Header extends Component {
                     }}
                   >
                     FEATURES
-                  </NavLink>
+                  </PLink>
                   <Link to="/landing#plans" style={{ color: "white" }}>
                     PRICING
                   </Link>
@@ -234,7 +236,7 @@ class Header extends Component {
         <Main>
           {this.props.loginSuccess ? (
             <HeaderWrapper>
-              <NavLink
+              <PLink
                 onClick={() => {
                   this.props.toggleMobileMenu(
                     this.props.mobileMenu,
@@ -243,7 +245,7 @@ class Header extends Component {
                 }}
               >
                 DASHBOARD
-              </NavLink>
+              </PLink>
               <TextLink to="/preferences">PREFERENCES</TextLink>
               {/* <Link to="/register">
                 <RegisterButton>REFER A FRIEND </RegisterButton>
