+++ src/components/Register/index.js
@@ -4,7 +4,8 @@ import { connect } from "react-redux";
 import styled from "styled-components";
 import { InputText } from "primereact/inputtext";
 import { Password } from "primereact/password";
-import { PrimaryButton, SpinnerWrapper } from "../Common";
+import PrimaryButton from "../Common/PrimaryButton";
+import { SpinnerWrapper } from "../Common";
 import { ProgressSpinner } from "primereact/progressspinner";
 import { Message } from "primereact/message";
 import { Lightbox } from "primereact/lightbox";
